SVT streamer can now accept requests for fractional mip levels. Smoothly changing the requested fractional mip level can currently lead to certain performance issues with DDC because additional small requests are made to stream in a few missing tiles at a time. To mitigate this, there is now a new cvar to discretize the computed streaming size. Also fixed a crash when unregistering a SVT from the streamer while it is currently playing back. Fixed by properly removing all FStreamingInstances belonging to the SVT to be removed.
#rb Patrick.Kelly
#rnx

[CL 32499747 by tim doerries in 5.4 branch]

+++ Engine/Source/Runtime/Engine/Private/SparseVolumeTexture/SparseVolumeTexture.cpp
@@ -939,7 +939,7 @@ USparseVolumeTextureFrame::USparseVolumeTextureFrame(const FObjectInitializer& O
 {
 }
 
-USparseVolumeTextureFrame* USparseVolumeTextureFrame::GetFrameAndIssueStreamingRequest(USparseVolumeTexture* SparseVolumeTexture, uint32 StreamingInstanceKey, float FrameRate, float FrameIndex, int32 MipLevel, bool bBlocking, bool bHasValidFrameRate)
+USparseVolumeTextureFrame* USparseVolumeTextureFrame::GetFrameAndIssueStreamingRequest(USparseVolumeTexture* SparseVolumeTexture, uint32 StreamingInstanceKey, float FrameRate, float FrameIndex, float MipLevel, bool bBlocking, bool bHasValidFrameRate)
 {
 	if (UStreamableSparseVolumeTexture* StreamableSVT = Cast<UStreamableSparseVolumeTexture>(SparseVolumeTexture))
 	{
